refactor: move instruction builders from dz_ledger_reader to dz_ledger_writer
