+++ crates/sentinel/src/client/mod.rs
@@ -1,4 +1,5 @@
 pub mod doublezero_ledger;
 pub mod dz_ledger_reader;
+pub mod dz_ledger_writer;
 pub mod solana;
 pub mod validator_metadata_reader;
